Moved struct to own module
